requirements.txt
asgiref==3.7.1
click==8.1.3
Django==4.2.1
django-phonenumber-field==7.1.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
filetype==1.2.0
h11==0.14.0
phonenumbers==8.13.11
Pillow==9.5.0
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.1
uvicorn==0.22.0
